chore: Move `workflow_response_converter` into `core/app/apps/common`
